src/requirements.txt
cloudflare>=2.8.15
PyYAML>=5.4.1
requests>=2.21.0
tldextract>=3.1.0